requirements.txt
requests==2.31.0
schedule==1.2.1
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.2
